feat: Add multiplatform support to coil-gif
Convert coil-gif from Android-only to Kotlin Multiplatform, enabling
animated GIF/WebP support on iOS, macOS, desktop, and web platforms.

Changes:
- Convert build.gradle.kts to multiplatform structure
- Add AnimatedSkiaImageDecoder for Skia-based platforms
- Add AnimatedImageDecoderFactory() expect/actual for platform-specific decoder selection
- Move common code to commonMain (decodeUtils, PixelOpacity, AnimatedTransformation)
- Add platform-specific ServiceLoader targets
- Update AnimatedTransformation to use coil3.Canvas typealias
- Add binary compatibility shims for moved Android functions
- Update documentation to use "animated image" instead of "animated drawable"

Based on PR #2594 with review comments addressed.
