+++ coil-gif/api/android/coil-gif.api
@@ -1,0 +1,118 @@
+public final class coil3/gif/AnimatedImageDecoder : coil3/decode/Decoder {
+	public static final field $stable I
+	public fun <init> (Lcoil3/decode/ImageSource;Lcoil3/request/Options;Z)V
+	public synthetic fun <init> (Lcoil3/decode/ImageSource;Lcoil3/request/Options;ZILkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public fun decode (Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
+}
+
+public final class coil3/gif/AnimatedImageDecoder$Factory : coil3/decode/Decoder$Factory {
+	public static final field $stable I
+	public fun <init> ()V
+	public fun <init> (Z)V
+	public synthetic fun <init> (ZILkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public fun create (Lcoil3/fetch/SourceFetchResult;Lcoil3/request/Options;Lcoil3/ImageLoader;)Lcoil3/decode/Decoder;
+}
+
+public final class coil3/gif/AnimatedImageDecoderFactory_androidKt {
+	public static final fun AnimatedImageDecoderFactory ()Lcoil3/decode/Decoder$Factory;
+}
+
+public abstract interface class coil3/gif/AnimatedTransformation {
+	public abstract fun transform (Landroid/graphics/Canvas;)Lcoil3/gif/PixelOpacity;
+}
+
+public final class coil3/gif/DecodeUtilsKt {
+	public static final fun isAnimatedHeif (Lcoil3/decode/DecodeUtils;Lokio/BufferedSource;)Z
+	public static final fun isAnimatedWebP (Lcoil3/decode/DecodeUtils;Lokio/BufferedSource;)Z
+	public static final fun isGif (Lcoil3/decode/DecodeUtils;Lokio/BufferedSource;)Z
+	public static final fun isHeif (Lcoil3/decode/DecodeUtils;Lokio/BufferedSource;)Z
+	public static final fun isWebP (Lcoil3/decode/DecodeUtils;Lokio/BufferedSource;)Z
+}
+
+public final class coil3/gif/GifDecoder : coil3/decode/Decoder {
+	public static final field $stable I
+	public fun <init> (Lcoil3/decode/ImageSource;Lcoil3/request/Options;Z)V
+	public synthetic fun <init> (Lcoil3/decode/ImageSource;Lcoil3/request/Options;ZILkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public fun decode (Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
+}
+
+public final class coil3/gif/GifDecoder$Factory : coil3/decode/Decoder$Factory {
+	public static final field $stable I
+	public fun <init> ()V
+	public fun <init> (Z)V
+	public synthetic fun <init> (ZILkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public fun create (Lcoil3/fetch/SourceFetchResult;Lcoil3/request/Options;Lcoil3/ImageLoader;)Lcoil3/decode/Decoder;
+	public final fun getEnforceMinimumFrameDelay ()Z
+}
+
+public final class coil3/gif/ImageRequestsKt {
+	public static final fun animatedTransformation (Lcoil3/request/ImageRequest$Builder;Lcoil3/gif/AnimatedTransformation;)Lcoil3/request/ImageRequest$Builder;
+	public static final fun getAnimatedTransformation (Lcoil3/Extras$Key$Companion;)Lcoil3/Extras$Key;
+	public static final fun getAnimatedTransformation (Lcoil3/request/ImageRequest;)Lcoil3/gif/AnimatedTransformation;
+	public static final fun getAnimatedTransformation (Lcoil3/request/Options;)Lcoil3/gif/AnimatedTransformation;
+	public static final fun getAnimationEndCallback (Lcoil3/Extras$Key$Companion;)Lcoil3/Extras$Key;
+	public static final fun getAnimationEndCallback (Lcoil3/request/ImageRequest;)Lkotlin/jvm/functions/Function0;
+	public static final fun getAnimationEndCallback (Lcoil3/request/Options;)Lkotlin/jvm/functions/Function0;
+	public static final fun getAnimationStartCallback (Lcoil3/Extras$Key$Companion;)Lcoil3/Extras$Key;
+	public static final fun getAnimationStartCallback (Lcoil3/request/ImageRequest;)Lkotlin/jvm/functions/Function0;
+	public static final fun getAnimationStartCallback (Lcoil3/request/Options;)Lkotlin/jvm/functions/Function0;
+	public static final fun onAnimationEnd (Lcoil3/request/ImageRequest$Builder;Lkotlin/jvm/functions/Function0;)Lcoil3/request/ImageRequest$Builder;
+	public static final fun onAnimationStart (Lcoil3/request/ImageRequest$Builder;Lkotlin/jvm/functions/Function0;)Lcoil3/request/ImageRequest$Builder;
+}
+
+public final class coil3/gif/ImageRequests_androidKt {
+	public static final synthetic fun animatedTransformation (Lcoil3/request/ImageRequest$Builder;Lcoil3/gif/AnimatedTransformation;)Lcoil3/request/ImageRequest$Builder;
+	public static final synthetic fun getAnimatedTransformation (Lcoil3/request/ImageRequest;)Lcoil3/gif/AnimatedTransformation;
+	public static final synthetic fun getAnimatedTransformation (Lcoil3/request/Options;)Lcoil3/gif/AnimatedTransformation;
+	public static final synthetic fun getAnimationEndCallback (Lcoil3/request/ImageRequest;)Lkotlin/jvm/functions/Function0;
+	public static final synthetic fun getAnimationEndCallback (Lcoil3/request/Options;)Lkotlin/jvm/functions/Function0;
+	public static final synthetic fun getAnimationStartCallback (Lcoil3/request/ImageRequest;)Lkotlin/jvm/functions/Function0;
+	public static final synthetic fun getAnimationStartCallback (Lcoil3/request/Options;)Lkotlin/jvm/functions/Function0;
+	public static final fun getRepeatCount (Lcoil3/Extras$Key$Companion;)Lcoil3/Extras$Key;
+	public static final fun getRepeatCount (Lcoil3/request/ImageRequest;)I
+	public static final fun getRepeatCount (Lcoil3/request/Options;)I
+	public static final synthetic fun onAnimationEnd (Lcoil3/request/ImageRequest$Builder;Lkotlin/jvm/functions/Function0;)Lcoil3/request/ImageRequest$Builder;
+	public static final synthetic fun onAnimationStart (Lcoil3/request/ImageRequest$Builder;Lkotlin/jvm/functions/Function0;)Lcoil3/request/ImageRequest$Builder;
+	public static final fun repeatCount (Lcoil3/request/ImageRequest$Builder;I)Lcoil3/request/ImageRequest$Builder;
+}
+
+public final class coil3/gif/MovieDrawable : android/graphics/drawable/Drawable, androidx/vectordrawable/graphics/drawable/Animatable2Compat {
+	public static final field $stable I
+	public static final field Companion Lcoil3/gif/MovieDrawable$Companion;
+	public static final field REPEAT_INFINITE I
+	public fun <init> (Landroid/graphics/Movie;)V
+	public fun <init> (Landroid/graphics/Movie;Landroid/graphics/Bitmap$Config;)V
+	public fun <init> (Landroid/graphics/Movie;Landroid/graphics/Bitmap$Config;Lcoil3/size/Scale;)V
+	public synthetic fun <init> (Landroid/graphics/Movie;Landroid/graphics/Bitmap$Config;Lcoil3/size/Scale;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public fun clearAnimationCallbacks ()V
+	public fun draw (Landroid/graphics/Canvas;)V
+	public final fun getAnimatedTransformation ()Lcoil3/gif/AnimatedTransformation;
+	public final fun getConfig ()Landroid/graphics/Bitmap$Config;
+	public fun getIntrinsicHeight ()I
+	public fun getIntrinsicWidth ()I
+	public fun getOpacity ()I
+	public final fun getRepeatCount ()I
+	public final fun getScale ()Lcoil3/size/Scale;
+	public fun isRunning ()Z
+	public fun registerAnimationCallback (Landroidx/vectordrawable/graphics/drawable/Animatable2Compat$AnimationCallback;)V
+	public fun setAlpha (I)V
+	public final fun setAnimatedTransformation (Lcoil3/gif/AnimatedTransformation;)V
+	public fun setColorFilter (Landroid/graphics/ColorFilter;)V
+	public final fun setRepeatCount (I)V
+	public fun start ()V
+	public fun stop ()V
+	public fun unregisterAnimationCallback (Landroidx/vectordrawable/graphics/drawable/Animatable2Compat$AnimationCallback;)Z
+}
+
+public final class coil3/gif/MovieDrawable$Companion {
+}
+
+public final class coil3/gif/PixelOpacity : java/lang/Enum {
+	public static final field OPAQUE Lcoil3/gif/PixelOpacity;
+	public static final field TRANSLUCENT Lcoil3/gif/PixelOpacity;
+	public static final field UNCHANGED Lcoil3/gif/PixelOpacity;
+	public static fun getEntries ()Lkotlin/enums/EnumEntries;
+	public static fun valueOf (Ljava/lang/String;)Lcoil3/gif/PixelOpacity;
+	public static fun values ()[Lcoil3/gif/PixelOpacity;
+}
+
