+++ coil-gif/src/androidInstrumentedTest/kotlin/coil3/gif/AnimatedTransformationTest.kt
