+++ coil-gif/src/androidInstrumentedTest/kotlin/coil3/gif/AnimationCallbacksTest.kt
