+++ coil-gif/src/androidMain/kotlin/coil3/gif/internal/FrameDelayRewritingSource.kt
