+++ coil-gif/src/androidUnitTest/kotlin/coil3/gif/GifDecoderServiceLoaderTest.kt
