+++ coil-gif/src/commonMain/kotlin/coil3/gif/PixelOpacity.kt
