+++ coil-gif/src/commonMain/kotlin/coil3/gif/internal/GifDecoderServiceLoaderTarget.kt
@@ -1,0 +1,8 @@
+package coil3.gif.internal
+
+import coil3.gif.AnimatedImageDecoderFactory
+import coil3.util.DecoderServiceLoaderTarget
+
+internal class GifDecoderServiceLoaderTarget : DecoderServiceLoaderTarget {
+    override fun factory() = AnimatedImageDecoderFactory()
+}
