+++ coil-gif/src/jvmCommonMain/resources/META-INF/services/coil3.util.DecoderServiceLoaderTarget
