+++ coil-gif/src/nonAndroidMain/kotlin/coil3/gif/AnimatedImageDecoderFactory.nonAndroid.kt
@@ -1,0 +1,6 @@
+package coil3.gif
+
+import coil3.decode.Decoder
+
+actual fun AnimatedImageDecoderFactory(): Decoder.Factory =
+    AnimatedSkiaImageDecoder.Factory()
